refactor: implement follow/unfollow toggle button with optimistic updates
